Add refresh token controller
