+++ auth/jwt/index.ts
@@ -1,3 +1,6 @@
 import configs from "./defaults";
+import router from "./routers";
 
 export const jwtConfigs = configs;
+
+export const jwtRouters = router;
