+++ auth/jwt/routers.ts
@@ -1,0 +1,6 @@
+import express from "express";
+
+const router = express.Router();
+router.route("/refresh").post();
+
+export default router;
