+++ configs/auth.config.ts
@@ -1,0 +1,6 @@
+const authConfig = {
+  JWT: {},
+  HTTP: {},
+};
+
+export default authConfig;
