+++ tsconfig.json
@@ -32,7 +32,7 @@
       // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
       // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
       "typeRoots": ["./types"],                                  /* Specify multiple folders that act like './node_modules/@types'. */
-      // "types": [],                                      /* Specify type package names to be included without being referenced in a source file. */
+      "types": ["./types", "./node_modules/@types"],                                      /* Specify type package names to be included without being referenced in a source file. */
       // "allowUmdGlobalAccess": true,                     /* Allow accessing UMD globals from modules. */
       // "moduleSuffixes": [],                             /* List of file name suffixes to search when resolving a module. */
       "resolveJsonModule": true,                        /* Enable importing .json files. */
